Feat: 자유게시판 API 구현
1. 새 코드 추가
- 자유 게시판 리스트 조회 API 구현
- 상세 조회 API 구현
- 상세 수정 API 구현
- 상세 삭제 API 구현

+++ backend/postings/views.py
@@ -1,52 +1,52 @@
 import json
 
-from django.views  import View
-from django.http   import JsonResponse
+from django.views import View
+from django.http  import JsonResponse
 
-from postings.models import FreeBoardPosting, FreeComment, FreeView
+from postings.models import NoticeBoardPosting, NoticeComment, NoticeView
 
-class FreeBoardListView(View):
+class NoticeBoardListView(View):
     """
     * @code writer 조현우
-    * @GET ("/postings/freeboards")
+    * @GET ("/postings/notices")
     *
     * @returns json
     """
     def get(self, request):
-        postings = FreeBoardPosting.objects.all()
+        postings = NoticeBoardPosting.objects.all()
 
         results = [{
             'id'      : posting.id,
             'title'   : posting.title,
             'context' : posting.context[:15],
-            'views'   : FreeView.objects.filter(free_board_posting=posting).count()
+            'views'   : NoticeView.objects.filter(notice_posting=posting).count()
         } for posting in postings ]
 
         return JsonResponse({'results' : results}, status = 200)
 
 
-class FreeBoardDetailView(View):
+class NoticeDetailView(View):
     """
     * @code writer 조현우
-    * @GET ("/postings/freeboards/detail")
+    * @GET ("/postings/notices/detail")
     *
     * @returns json
     """
     def get(self, request, posting_id):
         try:
-            posting = FreeBoardPosting.objects.get(id=posting_id)
-            comments = FreeComment.objects.filter(free_board_posting=posting)
+            posting = NoticeBoardPosting.objects.get(id=posting_id)
+            comments = NoticeComment.objects.filter(notice_posting=posting)
 
             result = {
                 'id'       : posting.id,
                 'title'    : posting.title,
                 'context'  : posting.context,
-                'views'    : FreeView.objects.filter(free_board_posting=posting).count(),
+                'views'    : NoticeView.objects.filter(notice_posting=posting).count(),
                 'comments' : [comment for comment in comments]
             }
 
             return JsonResponse({'result' : result}, status = 200)
-        except FreeBoardPosting.DoesNotExist:
+        except NoticeBoardPosting.DoesNotExist:
             return JsonResponse({'message' : 'POSTING_DOES_NOT_EXIST'}, status = 400)
 
     """
@@ -59,7 +59,10 @@ class FreeBoardDetailView(View):
         try:
             data    = json.loads(request.body)
 
-            posting, is_created = FreeBoardPosting.objects.update_or_create(
+            if user.level != 2:
+                return JsonResponse({'message' : 'NO_AUTHENTIFICATION'}, status = 403)
+
+            posting, is_created = NoticeBoardPosting.objects.update_or_create(
                 title    = data['title'],
                 context  = data['context'],
                 defaults = {
@@ -76,18 +79,23 @@ class FreeBoardDetailView(View):
 
     """
     * @code writer 조현우
-    * @DELETE ("/postings/freeboards/detail")
+    * @DELETE ("/postings/notices/detail")
     *
     * @returns json
     """
     def delete(self, request):
         try:
-            data    = json.loads(request)
-            posting = FreeBoardPosting.objects.get(id=data["posting_id"])
+            data = json.loads(request)
+            user = request.user
+
+            if user.level != 2:
+                return JsonResponse({'message' : 'NO_AUTHENTIFICATION'}, status = 403)
+
+            posting = NoticeBoardPosting.objects.get(id=data["posting_id"])
             posting.delete()
 
             return JsonResponse({'message' : 'DATA_DELETED'}, status = 204)
         except KeyError:
             return JsonResponse({'message' : 'KEY_ERROR'}, status = 400)
-        except FreeBoardPosting.DoesNotExist:
-            return JsonResponse({'message' : 'DATA_DOES_NOT_EXIST'}, status = 400)
+        except NoticeBoardPosting.DoesNotExist:
+            return JsonResponse({'message' : 'POSTING_DOES_NOT_EXIST'}, status = 400)
